+++ lib/punchblock/translator/freeswitch/call.rb
@@ -139,6 +139,9 @@ module Punchblock
           options[:origination_caller_id_number] = "'#{cid_number}'" if cid_number.present?
           options[:origination_caller_id_name] = "'#{cid_name}'" if cid_name.present?
           options[:originate_timeout] = dial_command.timeout/1000 if dial_command.timeout
+          dial_command.headers_hash.each_pair do |key, value|
+            options[key] = "'#{value}'"
+          end
           opts = options.inject([]) do |a, (k, v)|
             a << "#{k}=#{v}"
           end.join(',')